+++ app/client/authentication/check_authentication.js
@@ -5,13 +5,14 @@ export async function check_authentication(authenticationCode) {
 
     const savedAuthenticationCode = window.localStorage.getItem('authenticationCode')
     const authenticationField = document.getElementById('authentication-code');
+    const authenticationIntroField = document.getElementById('intro-authentication-code');
 
     if (!savedAuthenticationCode){ return }
 
     authenticationField.value = savedAuthenticationCode
+    authenticationIntroField.value = savedAuthenticationCode
 
     let authenticationStatus = await request_authentication(savedAuthenticationCode)
-    console.log('authenticationStatus: '+authenticationStatus)
 
     if (authenticationStatus){
         authenticated = authenticationStatus