+++ src/app.js
@@ -8,11 +8,14 @@ import { list_blacklist } from '../app/client/blacklist/list_blacklist.js'
 import { list_saved_searches } from '../app/client/saved_searches/list_saved_searches.js'
 import { load_event_listeners } from '../app/client/event_listeners/event_listeners.js'
 import { listener_authentication } from "../app/client/event_listeners/listener_authentication.js";
+import { listener_authentication_intro } from "../app/client/event_listeners/listener_authentication_intro.js";
 import { check_authentication } from "../app/client/authentication/check_authentication.js";
 import { intro_modal } from "../app/client/intro/intro_modal.js";
 
 await check_authentication()
 listener_authentication()
+listener_authentication_intro()
+intro_modal()
 
 // ┌─────────────────────────────────────┐
 // │                                     │
@@ -22,7 +25,6 @@ listener_authentication()
 console.log('what is authenticated: '+ authenticated)
 if (authenticated)
 {
-    intro_modal()
     onload_searches()
     load_inspector('positive')
     load_inspector('negative')